Fixing bugs in non-classic apps
Appbar layout was assuming classic, broken for online, browser, buildplane, etc.
I moved the menus & persistence title to an injected "spacer" slot for appbar.
Apps are now working, except for issues below.

The architecture for connectTrackballScene is finally right, using a special
controller action triggered in an effect from the Camera component.

Spurious trackball change events causing width=NaN is fixed, by making sure
that the callback only happens in the main LightedTrackballCanvas.

- no alert showing in apps
- open-change-save-reopen bug, losing parts of history
